+++ back-end/src/main/java/world/ludium/education/profile/dto/MyDetailAnnouncementDTO.java
@@ -8,6 +8,6 @@ public record MyDetailAnnouncementDTO(UUID detailId,
                                       Timestamp createAt,
                                       String status,
                                       UUID rewardToken,
-                                      Integer rewardAmount,
+                                      Double rewardAmount,
                                       String rewardClaimStatus) {
 }
